chore: relocate commons components to commons folder
