+++ src/components/index.js
@@ -1,4 +1,4 @@
-export { default as Confirmation } from './Confirmation';
-export { default as Footer } from './Footer';
-export { default as Navbar } from './Navbar';
-export { default as Skeleton } from './Skeleteon';
+export { default as Confirmation } from './commons/Confirmation';
+export { default as Footer } from './commons/Footer';
+export { default as Navbar } from './commons/Navbar';
+export { default as Skeleton } from './commons/Skeleteon';
